refactor: move token request to GetObjectId to improve performance
