+++ internal/clients/client.go
@@ -2,12 +2,7 @@ package clients
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
-	"errors"
-	"fmt"
 	"log"
-	"strings"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
@@ -28,6 +23,7 @@ type Client struct {
 	DataPlaneClient *DataPlaneClient
 
 	Account ResourceManagerAccount
+	Option  *Option
 }
 
 type Option struct {
@@ -47,6 +43,7 @@ type Option struct {
 func (client *Client) Build(ctx context.Context, o *Option) error {
 	client.StopContext = ctx
 	client.Features = o.Features
+	client.Option = o
 
 	azlog.SetListener(func(cls azlog.Event, msg string) {
 		log.Printf("[DEBUG] %s %s: %s\n", time.Now().Format(time.StampMicro), cls, msg)
@@ -91,16 +88,7 @@ func (client *Client) Build(ctx context.Context, o *Option) error {
 		"api-version",
 		"$skipToken",
 	}
-	tok, err := o.Cred.GetToken(client.StopContext, policy.TokenRequestOptions{
-		TenantID: o.TenantId,
-		Scopes:   []string{o.CloudCfg.Services[cloud.ResourceManager].Endpoint + "/.default"}})
-	if err != nil {
-		return fmt.Errorf("failed to get token to determine object id: %w", err)
-	}
-	cl, err := parseTokenClaims(tok.Token)
-	if err != nil {
-		return fmt.Errorf("failed to parse token claims to determine object id: %w", err)
-	}
+
 	resourceClient, err := NewResourceClient(o.Cred, &arm.ClientOptions{
 		ClientOptions: policy.ClientOptions{
 			Cloud: o.CloudCfg,
@@ -145,48 +133,7 @@ func (client *Client) Build(ctx context.Context, o *Option) error {
 	}
 	client.DataPlaneClient = dataPlaneClient
 
-	var oid string
-	if cl != nil && cl.ObjectId != "" {
-		oid = cl.ObjectId
-	}
-	client.Account = NewResourceManagerAccount(o.TenantId, o.SubscriptionId, oid)
+	client.Account = NewResourceManagerAccount(client)
 
 	return nil
-}
-
-func parseTokenClaims(token string) (*tokenClaims, error) {
-	// Parse the token to get the claims
-	parts := strings.Split(token, ".")
-	if len(parts) != 3 {
-		return nil, errors.New("parseTokenClaims: token does not have 3 parts")
-	}
-	decoded, err := base64.RawURLEncoding.DecodeString(parts[1])
-	if err != nil {
-		return nil, fmt.Errorf("parseTokenClaims: error decoding token: %s", err)
-	}
-	var claims tokenClaims
-	err = json.Unmarshal(decoded, &claims)
-	if err != nil {
-		return nil, fmt.Errorf("parseTokenClaims: error unmarshalling claims: %w", err)
-	}
-	return &claims, nil
-}
-
-type tokenClaims struct {
-	Audience          string   `json:"aud"`
-	Expires           int64    `json:"exp"`
-	IssuedAt          int64    `json:"iat"`
-	Issuer            string   `json:"iss"`
-	IdentityProvider  string   `json:"idp"`
-	ObjectId          string   `json:"oid"`
-	Roles             []string `json:"roles"`
-	Scopes            string   `json:"scp"`
-	Subject           string   `json:"sub"`
-	TenantRegionScope string   `json:"tenant_region_scope"`
-	TenantId          string   `json:"tid"`
-	Version           string   `json:"ver"`
-
-	AppDisplayName string `json:"app_displayname,omitempty"`
-	AppId          string `json:"appid,omitempty"`
-	IdType         string `json:"idtyp,omitempty"`
 }
